Refactor query test cases code
