+++ mp2-v1/tests/integrated_tests.rs
@@ -17,10 +17,9 @@ use anyhow::{Context, Result};
 use common::{
     cases::{
         indexing::{ChangeType, TreeFactory, UpdateType},
-        query::{
+        query::{aggregated_queries::{
             test_query, GlobalCircuitInput, QueryCircuitInput, RevelationCircuitInput,
-            MAX_NUM_PLACEHOLDERS,
-        },
+        },MAX_NUM_PLACEHOLDERS}
     },
     context::{self, ParamsType, TestContextConfig},
     proof_storage::{ProofKV, DEFAULT_PROOF_STORE_FOLDER},
